xcm: reorganized and converted xcm_config into modules
