+++ runtime/src/configs/xcm_config/mod.rs
@@ -1,0 +1,6 @@
+mod asset_matcher;
+mod origin_filters;
+mod weight_trader;
+mod config;
+
+pub use config::*;
